+++ src/Framework/Bootstrap/SetupGacela.php
@@ -6,14 +6,11 @@ namespace Gacela\Framework\Bootstrap;
 
 use Closure;
 use Gacela\Framework\ClassResolver\Cache\GacelaFileCache;
+use Gacela\Framework\Config\GacelaConfigBuilder\AppConfigBuilder;
 use Gacela\Framework\Config\GacelaConfigBuilder\BindingsBuilder;
-use Gacela\Framework\Config\GacelaConfigBuilder\ConfigBuilder;
 use Gacela\Framework\Config\GacelaConfigBuilder\SuffixTypesBuilder;
 use Gacela\Framework\Container\Container;
-use Gacela\Framework\Event\Dispatcher\ConfigurableEventDispatcher;
 use Gacela\Framework\Event\Dispatcher\EventDispatcherInterface;
-use Gacela\Framework\Event\Dispatcher\NullEventDispatcher;
-
 use RuntimeException;
 
 use function is_callable;
@@ -31,7 +28,7 @@ final class SetupGacela extends AbstractSetupGacela
     public const configKeyValues = 'configKeyValues';
     public const servicesToExtend = 'servicesToExtend';
     public const plugins = 'plugins';
-    public const extendConfig = 'extendConfig';
+    public const gacelaConfigsToExtend = 'gacelaConfigsToExtend';
 
     private const DEFAULT_ARE_EVENT_LISTENERS_ENABLED = true;
     private const DEFAULT_SHOULD_RESET_IN_MEMORY_CACHE = false;
@@ -42,11 +39,11 @@ final class SetupGacela extends AbstractSetupGacela
     private const DEFAULT_GENERIC_LISTENERS = [];
     private const DEFAULT_SPECIFIC_LISTENERS = [];
     private const DEFAULT_SERVICES_TO_EXTEND = [];
-    private const DEFAULT_EXTEND_CONFIG = [];
+    private const DEFAULT_GACELA_CONFIGS_TO_EXTEND = [];
     private const DEFAULT_PLUGINS = [];
 
-    /** @var callable(ConfigBuilder):void */
-    private $configFn;
+    /** @var callable(AppConfigBuilder):void */
+    private $appConfigFn;
 
     /** @var callable(BindingsBuilder,array<string,mixed>):void */
     private $bindingsFn;
@@ -57,7 +54,7 @@ final class SetupGacela extends AbstractSetupGacela
     /** @var ?array<string,class-string|object|callable> */
     private ?array $externalServices = null;
 
-    private ?ConfigBuilder $configBuilder = null;
+    private ?AppConfigBuilder $appConfigBuilder = null;
 
     private ?SuffixTypesBuilder $suffixTypesBuilder = null;
 
@@ -92,9 +89,9 @@ final class SetupGacela extends AbstractSetupGacela
     private ?array $servicesToExtend = null;
 
     /** @var ?list<class-string> */
-    private ?array $extendConfig = null;
+    private ?array $gacelaConfigsToExtend = null;
 
-    /** @var ?list<class-string> */
+    /** @var ?list<class-string|callable> */
     private ?array $plugins = null;
 
     public function __construct()
@@ -102,7 +99,7 @@ final class SetupGacela extends AbstractSetupGacela
         $emptyFn = static function (): void {
         };
 
-        $this->configFn = $emptyFn;
+        $this->appConfigFn = $emptyFn;
         $this->bindingsFn = $emptyFn;
         $this->suffixTypesFn = $emptyFn;
     }
@@ -140,9 +137,9 @@ final class SetupGacela extends AbstractSetupGacela
 
         return (new self())
             ->setExternalServices($build['external-services'])
-            ->setConfigBuilder($build['config-builder'])
+            ->setAppConfigBuilder($build['app-config-builder'])
             ->setSuffixTypesBuilder($build['suffix-types-builder'])
-            ->setBindingsBuilder($build['mapping-interfaces-builder'])
+            ->setBindingsBuilder($build['bindings-builder'])
             ->setShouldResetInMemoryCache($build['should-reset-in-memory-cache'])
             ->setFileCacheEnabled($build['file-cache-enabled'])
             ->setFileCacheDirectory($build['file-cache-directory'])
@@ -151,9 +148,9 @@ final class SetupGacela extends AbstractSetupGacela
             ->setAreEventListenersEnabled($build['are-event-listeners-enabled'])
             ->setGenericListeners($build['generic-listeners'])
             ->setSpecificListeners($build['specific-listeners'])
-            ->setExtendConfig($build['before-config'])
-            ->setPlugins($build['after-plugins'])
-            ->setServicesToExtend($build['services-to-extend']);
+            ->setGacelaConfigsToExtend($build['gacela-configs-to-extend'])
+            ->setPlugins($build['plugins'])
+            ->setServicesToExtend($build['instances-to-extend']);
     }
 
     /**
@@ -167,9 +164,9 @@ final class SetupGacela extends AbstractSetupGacela
         return $this;
     }
 
-    public function setConfigBuilder(ConfigBuilder $builder): self
+    public function setAppConfigBuilder(AppConfigBuilder $builder): self
     {
-        $this->configBuilder = $builder;
+        $this->appConfigBuilder = $builder;
 
         return $this;
     }
@@ -189,24 +186,24 @@ final class SetupGacela extends AbstractSetupGacela
     }
 
     /**
-     * @param callable(ConfigBuilder):void $callable
+     * @param callable(AppConfigBuilder):void $callable
      */
-    public function setConfigFn(callable $callable): self
+    public function setAppConfigFn(callable $callable): self
     {
-        $this->configFn = $callable;
+        $this->appConfigFn = $callable;
 
         return $this;
     }
 
-    public function buildConfig(ConfigBuilder $builder): ConfigBuilder
+    public function buildAppConfig(AppConfigBuilder $builder): AppConfigBuilder
     {
-        $builder = parent::buildConfig($builder);
+        $builder = parent::buildAppConfig($builder);
 
-        if ($this->configBuilder) {
-            $builder = $this->configBuilder;
+        if ($this->appConfigBuilder) {
+            $builder = $this->appConfigBuilder;
         }
 
-        ($this->configFn)($builder);
+        ($this->appConfigFn)($builder);
 
         return $builder;
     }
@@ -341,24 +338,7 @@ final class SetupGacela extends AbstractSetupGacela
 
     public function getEventDispatcher(): EventDispatcherInterface
     {
-        if ($this->eventDispatcher !== null) {
-            return $this->eventDispatcher;
-        }
-
-        if ($this->canCreateEventDispatcher()) {
-            $this->eventDispatcher = new ConfigurableEventDispatcher();
-            $this->eventDispatcher->registerGenericListeners($this->genericListeners ?? []);
-
-            foreach ($this->specificListeners ?? [] as $event => $listeners) {
-                foreach ($listeners as $callable) {
-                    $this->eventDispatcher->registerSpecificListener($event, $callable);
-                }
-            }
-        } else {
-            $this->eventDispatcher = new NullEventDispatcher();
-        }
-
-        return $this->eventDispatcher;
+        return $this->eventDispatcher ??= SetupEventDispatcher::getDispatcher($this);
     }
 
     /**
@@ -459,13 +439,13 @@ final class SetupGacela extends AbstractSetupGacela
     /**
      * @param list<class-string> $list
      */
-    public function combineExtendConfig(array $list): void
+    public function combineGacelaConfigsToExtend(array $list): void
     {
-        $this->setExtendConfig(array_merge($this->extendConfig ?? [], $list));
+        $this->setGacelaConfigsToExtend(array_merge($this->gacelaConfigsToExtend ?? [], $list));
     }
 
     /**
-     * @param list<class-string> $list
+     * @param list<class-string|callable> $list
      */
     public function combinePlugins(array $list): void
     {
@@ -475,31 +455,33 @@ final class SetupGacela extends AbstractSetupGacela
     /**
      * @return list<class-string>
      */
-    public function getExtendConfig(): array
+    public function getGacelaConfigsToExtend(): array
     {
-        return (array)$this->extendConfig;
+        return (array)$this->gacelaConfigsToExtend;
     }
 
     /**
-     * @return list<class-string>
+     * @return list<class-string|callable>
      */
     public function getPlugins(): array
     {
         return (array)$this->plugins;
     }
 
-    private static function runExtendConfig(GacelaConfig $config): void
+    private static function runExtendConfig(GacelaConfig $gacelaConfig): void
     {
-        $plugins = $config->build()['before-config'] ?? [];
+        $configsToExtend = $gacelaConfig->build()['gacela-configs-to-extend'] ?? [];
 
-        if ($plugins === []) {
+        if ($configsToExtend === []) {
             return;
         }
 
-        foreach ($plugins as $pluginName) {
-            /** @var callable $plugin */
-            $plugin = Container::create($pluginName);
-            $plugin($config);
+        $container = new Container();
+
+        foreach ($configsToExtend as $className) {
+            /** @var callable $configToExtend */
+            $configToExtend = $container->get($className);
+            $configToExtend($gacelaConfig);
         }
     }
 
@@ -540,16 +522,16 @@ final class SetupGacela extends AbstractSetupGacela
     /**
      * @param ?list<class-string> $list
      */
-    private function setExtendConfig(?array $list): self
+    private function setGacelaConfigsToExtend(?array $list): self
     {
-        $this->markPropertyChanged(self::extendConfig, $list);
-        $this->extendConfig = $list ?? self::DEFAULT_EXTEND_CONFIG;
+        $this->markPropertyChanged(self::gacelaConfigsToExtend, $list);
+        $this->gacelaConfigsToExtend = $list ?? self::DEFAULT_GACELA_CONFIGS_TO_EXTEND;
 
         return $this;
     }
 
     /**
-     * @param ?list<class-string> $list
+     * @param ?list<class-string|callable> $list
      */
     private function setPlugins(?array $list): self
     {