+++ src/Framework/Gacela.php
@@ -18,10 +18,16 @@ use Gacela\Framework\Config\ConfigFactory;
 use Gacela\Framework\Container\Container;
 use Gacela\Framework\Container\Locator;
 use Gacela\Framework\DocBlockResolver\DocBlockResolverCache;
+use Gacela\Framework\Exception\GacelaNotBootstrappedException;
+
+use function is_string;
 
 final class Gacela
 {
+    private const GACELA_PHP_FILENAME = 'gacela.php';
+
     private static ?Container $mainContainer = null;
+    private static ?string $appRootDir = null;
 
     /**
      * Define the entry point of Gacela.
@@ -30,22 +36,13 @@ final class Gacela
      */
     public static function bootstrap(string $appRootDir, Closure $configFn = null): void
     {
+        self::$appRootDir = $appRootDir;
         self::$mainContainer = null;
 
-        $setup = self::processConfigFnIntoSetup($appRootDir, $configFn);
+        $setup = self::processConfigFnIntoSetup($configFn);
 
         if ($setup->shouldResetInMemoryCache()) {
-            AbstractFacade::resetCache();
-            AnonymousGlobal::resetCache();
-            AbstractFactory::resetCache();
-            GacelaFileCache::resetCache();
-            DocBlockResolverCache::resetCache();
-            ClassResolverCache::resetCache();
-            InMemoryCache::resetCache();
-            AbstractClassResolver::resetCache();
-            ConfigFactory::resetCache();
-            Config::resetInstance();
-            Locator::resetInstance();
+            self::resetCache();
         }
 
         $config = Config::createWithSetup($setup);
@@ -68,15 +65,26 @@ final class Gacela
     }
 
     /**
+     * Get the application root dir set when bootstrapping gacela
+     */
+    public static function rootDir(): string
+    {
+        if (self::$appRootDir === null) {
+            throw new GacelaNotBootstrappedException();
+        }
+        return self::$appRootDir;
+    }
+
+    /**
      * @param null|Closure(GacelaConfig):void $configFn
      */
-    private static function processConfigFnIntoSetup(string $appRootDir, Closure $configFn = null): SetupGacelaInterface
+    private static function processConfigFnIntoSetup(Closure $configFn = null): SetupGacelaInterface
     {
         if ($configFn !== null) {
             return SetupGacela::fromCallable($configFn);
         }
 
-        $gacelaFilePath = $appRootDir . '/gacela.php';
+        $gacelaFilePath = sprintf('%s%s%s', self::rootDir(), DIRECTORY_SEPARATOR, self::GACELA_PHP_FILENAME);
 
         if (is_file($gacelaFilePath)) {
             return SetupGacela::fromFile($gacelaFilePath);
@@ -85,16 +93,34 @@ final class Gacela
         return new SetupGacela();
     }
 
+    private static function resetCache(): void
+    {
+        AbstractFacade::resetCache();
+        AnonymousGlobal::resetCache();
+        AbstractFactory::resetCache();
+        GacelaFileCache::resetCache();
+        DocBlockResolverCache::resetCache();
+        ClassResolverCache::resetCache();
+        InMemoryCache::resetCache();
+        AbstractClassResolver::resetCache();
+        ConfigFactory::resetCache();
+        Config::resetInstance();
+        Locator::resetInstance();
+    }
+
     private static function runPlugins(Config $config): void
     {
         self::$mainContainer = Container::withConfig($config);
 
         $plugins = $config->getSetupGacela()->getPlugins();
 
-        foreach ($plugins as $pluginName) {
-            /** @var callable $plugin */
-            $plugin = self::$mainContainer->get($pluginName);
-            $plugin();
+        foreach ($plugins as $plugin) {
+            /** @var callable $current */
+            $current = is_string($plugin)
+                ? self::$mainContainer->get($plugin)
+                : $plugin;
+
+            self::$mainContainer->resolve($current);
         }
     }
 }